Remove circular reference by moving KnownCosmosDbCmkSubStatusValue into Core.
