+++ src/Microsoft.Health.Fhir.Core/Exceptions/KnownCosmosDbCmkSubStatusValue.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-namespace Microsoft.Health.CosmosDb.Features.Storage
+namespace Microsoft.Health.Fhir.Core.Exceptions
 {
     /// <summary>
     /// Cosmos DB customer-managed key (CMK) sub status values
